+++ probe-rs/src/config/sequences/nrf.rs
@@ -28,8 +28,8 @@ pub trait Nrf: Sync + Send + Debug {
     fn has_network_core(&self) -> bool;
 }
 
-const ERASEALL: u8 = 0x04;
-const ERASEALLSTATUS: u8 = 0x08;
+const ERASEALL: u16 = 0x04;
+const ERASEALLSTATUS: u16 = 0x08;
 
 const APPLICATION_SPU_PERIPH_PERM: u64 = 0x50003800;
